see #1344: fix sitemap link

+++ src/admin/partials/wordlift-admin-settings-videoobject-settings-page.php
@@ -82,11 +82,11 @@ if ( isset( $_POST['submit'] ) ) {
     <p>
 		<?php _e( 'The Video Sitemap works like any other XML Sitemap. Search engines will use it to display rich snippets in result pages.' ); ?>
     </p>
-	<?php $is_checked = get_option( '_wl_video_sitemap_generation', false ) ? 'checked' : '' ?>
+	<?php $is_sitemap_enabled = get_option( '_wl_video_sitemap_generation', false ) ? 'checked' : '' ?>
     <p> <?php _e( 'Enable Video Sitemap' ); ?>
-        <input type="checkbox" name="wl_enable_video_sitemap" value="1" <?php echo $is_checked; ?> ></p>
+        <input type="checkbox" name="wl_enable_video_sitemap" value="1" <?php echo $is_sitemap_enabled; ?> ></p>
     <p> <?php
-		if ( Video_Sitemap::is_sitemap_already_generated() ) {
+		if ( $is_sitemap_enabled && Video_Sitemap::is_sitemap_already_generated() ) {
 			$sitemap_link = get_home_url( null, 'wl-video-sitemap.xml' );
 			printf( __( 'Here is <a href="%s">link</a> to your Video Sitemap. Add it now, to Google Search Console.', 'wordlift' ), $sitemap_link );
 		}
